doctracer/requirements.txt
fastapi
uvicorn
strawberry-graphql[fastapi]
google-generativeai
python-multipart
python-dotenv
pypdf
